split rust reports into a separate module (I think that's the right word?)
